requirements.txt
argon2-cffi==23.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
pytest-mock==3.14.0
bleach==6.3.0
cryptography==41.0.7
fastapi==0.135.3
pydantic[email]==2.13.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-socketio==5.16.1
SQLAlchemy==2.0.49
starlette==1.0.0
uvicorn==0.44.0
dynaconf==3.2.13
psycopg2-binary==2.9.11
alembic==1.16.0